Moves analytics to their own directory
